Fix some return values in the language server (#2134)
* Fix return value of LSP shutdown

It is specified as a request that returns null. Returning void causes
the language server to crash.

* Fix empty responses of LSP textDocument/definition

It is specified to return either a Location, an array of Locations, or
null. Hovers are unrelated.

* Fix responses of LSP textDocument/hover

It is specified to either return Hover or null. Psalm could previously
return a success response containing a MarkedString, but with a null
value. MarkedString is not allowed to contain null.