Draft: Do not use snippets when LSP client does not support them
Returning completion of kind SNIPPET is only allowed when the client has
advertised support. This changes the method completion code to only
return a snippet when the client supports them, and when the method has
arguments. Otherwise, a plaintext completion is returned.

When the function has no arguments, this now also returns a completion
without parenthesis, for consistency with other language server
implementation.

TODO: This does not actually check the client capabilities yet, but only
implements the output handling.

This fixes #4139.