Added test to keep Union commutative
Union should be a commutative operation, such that `A|B = B|A`. This
seems to be the case with types parsed from strings, but not for unions
created directly in the code.

E.g.
```
new Union([Atomic::create('int'), Atomic::create('positive-int')])
```
currently results in `positive-int`, even though int should subsume it.

Refs vimeo/psalm#5342