Pretty print arrays in LSP hover
Currently hovering keyed array definitions is very unreadable. Opening this PR to discuss a good solution. I attached some code, but it's not great. The idea is to have this:

![](https://joehoyle-captured.s3.amazonaws.com/Screen-Shot-2021-12-05-07-54-21.99-7nfVN8oFX9Q3ErfpKygsOJDUTYfs2n8o7MDkW0iINRQ9tIHgpI5YBf5RnJhBu63PLvPleQPLNP14HeQZbkbLB9QzdPzZomzh05dA.png)

Look like this:

![](https://joehoyle-captured.s3.amazonaws.com/Screen-Shot-2021-12-05-07-55-15.02-4fRbE8xThSsDSIZjeHGctkco4ZrfBsPbwn40SbySdiDkwqa2Px0lUU8acnxJdcLhj7DMuEeqsCOqztuIGFvv82e8LqWVbjjYJEsT.png)

Somewhat inspired by how the TypeScript LSP shows object literals:

![](https://joehoyle-captured.s3.amazonaws.com/Screen-Shot-2021-12-05-07-56-41.86-wolTT97NaK3OCDZm9PJIucBaCXv4O3ep4uIiEseOfXGuAasveoJXHOPmC5uxszd72sz3BMTH92OR6VbH2xbPdbwuopU6cvxqUqwA.png)

---

The implementation here is not that good, ideally I think TKeyedArray should have a `prettyPrint` method perhaps. Also, there's other things to consider, like how to show `array{}` printing when then are embedded in `<list>` or `array<>` etc, where line breaks would go, etc.

This "re-parsing" of the type also is not great, this is because the file-maps (specifically the reference map) flattens type to their string representation. Because I _think_ that's potentially done in other threads, I'm assuming that's why.

Then there's the question of syntax highlighting for `array{}` which would help a lot with readability. The issue ultimately is that we rely on the LSP client to syntax highlght based off `<?php $type` code samples. Given that `array{}` is not valid PHP, then we don't send it as `<?php` prefixed, hence no formatting. One optino could be to actually encode it as `<?php [ "key" => "value" ]` style, though I'm not sure how something like `<?php [ "foo" => string ]` would render.