Reorganize pushes to `psalm/phar`
Previously we built phar and pushed it:

 1. To `psalm/phar:master` (always)
 2. To `psalm/phar:$tag` (for tagged releases)

However it's entirely possible to tag branches that diverged from master
(like when we do a patch release for a legacy version). In this case our
push to `psalm/phar:master` was rejected and script failed. As a result,
`psalm/phar` was missing the tag (release).

Now we will either:
 * push to `psalm/phar:master` (if the build was for `vimeo/psalm:master`)
 * or push to `psalm/phar:$tag` (if it's a tagged release)