Part 1. Add variable method_exists
Deep processing of mixins is planned to be done through recursive calls
to the `handleMixins` method.

To realise this, we need this method to return information that it has
managed to find the desired method.

Therefore, the `method_exists` variable was declared to make it clear
that the method was found.

The existing `naive_method_exists` variable is not suitable for us,
because it is not set to `true` if the method is magic (pseudo).