Remove usage of E_STRICT
This constant is deprecated in PHP 8.4. All E_STRICT warnings were
converted to E_NOTICE in PHP 8.0.

Psalm 5 still supports PHP 7.4, so _technically_ this constant could
still be useful for people running it on that version, but I suspect
most development and testing happens against 8.0+. It doesn't seem worth
more complicated version detection in this code to me.